requirements.txt
pandas
fpdf
GitPython